move getEditTransactionStatus and getMinFees in coin-evm
