+++ libs/coin-evm/src/getTransactionStatus.ts
@@ -1,31 +1,35 @@
-import {
-  NotEnoughGas,
-  FeeNotLoaded,
-  InvalidAddress,
-  ETHAddressNonEIP,
-  RecipientRequired,
-  AmountRequired,
-  NotEnoughBalance,
-  GasLessThanEstimate,
-  PriorityFeeTooLow,
-  PriorityFeeTooHigh,
-  PriorityFeeHigherThanMaxFee,
-  MaxFeeTooLow,
-  NotEnoughBalanceInParentAccount,
-  FeeTooHigh,
-} from "@ledgerhq/errors";
-import { ethers } from "ethers";
-import BigNumber from "bignumber.js";
-import { Account, AccountBridge, SubAccount } from "@ledgerhq/types-live";
 import { findSubAccountById } from "@ledgerhq/coin-framework/account/index";
-import { eip1559TransactionHasFees, getEstimatedFees, legacyTransactionHasFees } from "./logic";
 import {
+  AmountRequired,
+  ETHAddressNonEIP,
+  FeeNotLoaded,
+  FeeTooHigh,
+  GasLessThanEstimate,
+  InvalidAddress,
+  MaxFeeTooLow,
+  NotEnoughBalance,
+  NotEnoughBalanceInParentAccount,
+  NotEnoughGas,
+  PriorityFeeHigherThanMaxFee,
+  PriorityFeeTooHigh,
+  PriorityFeeTooLow,
+  RecipientRequired,
+  ReplacementTransactionUnderpriced,
+} from "@ledgerhq/errors";
+import { Account, AccountBridge, SubAccount, TransactionStatusCommon } from "@ledgerhq/types-live";
+import BigNumber from "bignumber.js";
+import { ethers } from "ethers";
+import { NotEnoughNftOwned, NotOwnedNft, QuantityNeedsToBePositive } from "./errors";
+import { getMinEip1559Fees, getMinLegacyFees } from "./getMinEditTransactionFees";
+import { eip1559TransactionHasFees, getEstimatedFees, legacyTransactionHasFees } from "./logic";
+import { DEFAULT_GAS_LIMIT } from "./transaction";
+import {
+  EditType,
+  Transaction as EvmTransaction,
   EvmTransactionEIP1559,
   EvmTransactionLegacy,
-  Transaction as EvmTransaction,
+  TransactionStatus,
 } from "./types";
-import { NotEnoughNftOwned, NotOwnedNft, QuantityNeedsToBePositive } from "./errors";
-import { DEFAULT_GAS_LIMIT } from "./transaction";
 
 type ValidatedTransactionFields =
   | "recipient"
@@ -284,6 +288,156 @@ export const getTransactionStatus: AccountBridge<EvmTransaction>["getTransaction
     amount,
     totalSpent,
   };
+};
+
+type EditTransactionValidatedTransactionFields = "replacementTransactionUnderpriced";
+type EditTransactionValidationIssues = Partial<
+  Record<EditTransactionValidatedTransactionFields, Error>
+>;
+
+/**
+ * Validate an edited transaction and returns related errors and warnings
+ * Makes sure the updated fees are at least 10% higher than the original fees
+ * 10% isn't defined in the protocol, it's just how most nodes & miners implement it
+ * If the new transaction fees are less than 10% higher than the original fees,
+ * the transaction will be rejected by the network with a
+ * "replacement transaction underpriced" error
+ */
+const validateEditTransaction = ({
+  editType,
+  transaction,
+  transactionToUpdate,
+}: {
+  transaction: EvmTransaction;
+  transactionToUpdate: EvmTransaction;
+  editType?: EditType;
+}): {
+  errors: TransactionStatusCommon["errors"];
+  warnings: TransactionStatusCommon["warnings"];
+} => {
+  const errors: EditTransactionValidationIssues = {};
+  const warnings: EditTransactionValidationIssues = {};
+
+  if (!editType) {
+    return {
+      errors: {},
+      warnings: {},
+    };
+  }
+
+  const { maxFeePerGas, maxPriorityFeePerGas, gasPrice } = transactionToUpdate;
+  const {
+    maxFeePerGas: newMaxFeePerGas,
+    maxPriorityFeePerGas: newMaxPriorityFeePerGas,
+    gasPrice: newGasPrice,
+  } = transaction;
+
+  if (transaction.type === 2) {
+    if (!maxFeePerGas || !maxPriorityFeePerGas) {
+      throw new Error("maxFeePerGas and maxPriorityFeePerGas are required");
+    }
+
+    if (!newMaxFeePerGas || !newMaxPriorityFeePerGas) {
+      // This should already be handled in getTransactionStatus
+      return {
+        errors: {},
+        warnings: {},
+      };
+    }
+
+    const { maxFeePerGas: minNewMaxFeePerGas, maxPriorityFeePerGas: minNewMaxPriorityFeePerGas } =
+      getMinEip1559Fees({ maxFeePerGas, maxPriorityFeePerGas });
+
+    if (
+      newMaxFeePerGas.isLessThan(minNewMaxFeePerGas) ||
+      newMaxPriorityFeePerGas.isLessThan(minNewMaxPriorityFeePerGas)
+    ) {
+      errors.replacementTransactionUnderpriced = new ReplacementTransactionUnderpriced();
+
+      return {
+        errors,
+        warnings,
+      };
+    }
+  } else {
+    if (!gasPrice) {
+      throw new Error("gasPrice is required");
+    }
+
+    if (!newGasPrice) {
+      // This should already be handled in getTransactionStatus
+      return {
+        errors: {},
+        warnings: {},
+      };
+    }
+
+    const { gasPrice: minNewGasPrice } = getMinLegacyFees({ gasPrice });
+
+    if (newGasPrice.isLessThan(minNewGasPrice)) {
+      errors.replacementTransactionUnderpriced = new ReplacementTransactionUnderpriced();
+
+      return {
+        errors,
+        warnings,
+      };
+    }
+  }
+
+  return {
+    errors: {},
+    warnings: {},
+  };
+};
+
+const ERROR_NOT_ENOUGH_NFT_OWNED = new NotEnoughNftOwned();
+const ERROR_NOT_OWNED_NFT = new NotOwnedNft();
+const ERROR_AMOUNT_REQUIRED = new AmountRequired();
+
+/**
+ * This function is used to update the status of an edited transaction
+ */
+export const getEditTransactionStatus = ({
+  transaction,
+  transactionToUpdate,
+  status,
+  editType,
+}: {
+  transaction: EvmTransaction;
+  transactionToUpdate: EvmTransaction;
+  status: TransactionStatus;
+  editType?: EditType;
+}): TransactionStatus => {
+  const { errors: editTxErrors } = validateEditTransaction({
+    editType,
+    transaction,
+    transactionToUpdate,
+  });
+
+  const errors: Record<string, Error> = { ...status.errors, ...editTxErrors };
+
+  // copy errors object to avoid mutating the original one
+  const updatedErrors: Record<string, Error> = { ...errors };
+
+  // discard "AmountRequired" (for cancel and speedup since one can decide to speedup a cancel)
+  // exclude "NotOwnedNft" and "NotEnoughNftOwned" error if it's a nft operation
+  // since nfts being sent are not owned by the user anymore (removed from the account)
+  // cf. this (as of 14-09-2023) https://github.com/LedgerHQ/ledger-live/blob/73217c9aa93c901d9c8d2067dcc7090243d35d35/libs/coin-evm/src/synchronization.ts#L111-L120
+  if (
+    updatedErrors.amount &&
+    (updatedErrors.amount.name === ERROR_NOT_OWNED_NFT.name ||
+      updatedErrors.amount.name === ERROR_NOT_ENOUGH_NFT_OWNED.name ||
+      updatedErrors.amount.name === ERROR_AMOUNT_REQUIRED.name)
+  ) {
+    delete updatedErrors.amount;
+  }
+
+  const updatedStatus: TransactionStatus = {
+    ...status,
+    errors: updatedErrors,
+  };
+
+  return updatedStatus;
 };
 
 export default getTransactionStatus;
