+++ libs/coin-evm/src/types/editTransaction.ts
@@ -1,0 +1,2 @@
+// Possible types of "replace by fee" (aka "edit transaction") operations:
+export type EditType = "cancel" | "speedup";
