+++ libs/coin-evm/src/types/index.ts
@@ -1,5 +1,6 @@
 /* istanbul ignore file: pure exports, bridge tested by live-common with bridge.integration.test.ts */
 
-export * from "./ledger";
+export * from "./editTransaction";
 export * from "./etherscan";
+export * from "./ledger";
 export * from "./transaction";
