--- libs/ledger-live-common/src/families/evm/editTransaction/types.ts
@@ -1,1 +1,0 @@
-export type EditType = "cancel" | "speedup";
